Fix app launch crashes
