+++ WeatherApp/App delegate/AppDelegate.swift
@@ -3,28 +3,32 @@ import GooglePlaces
 
 @UIApplicationMain
 class AppDelegate: UIResponder, UIApplicationDelegate {
-
-
+    
     var window: UIWindow?
 
     func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
         
         GMSPlacesClient.provideAPIKey("AIzaSyA7a9OOtHJ9Ol_yodYPZzs8OO9WYS9vD9Y")
         
-//        let storageManager = StorageManager.shared
-//
-//        if storageManager.containsCoordinates {
-//            if let coordinates = storageManager.storedCoordinates {
-//                let geoDataFinder = GeoDataFinder()
-//                geoDataFinder.delegate = self
-//                geoDataFinder.findBySpecified(coordinates: coordinates)
-//            }
-//        } else {
-//            let storyboard = UIStoryboard(name: "Finder", bundle: Bundle.main)
-//            let vc = storyboard.instantiateInitialViewController() as! FinderViewController
-//            UIApplication.shared.windows.first?.rootViewController = vc
-//            UIApplication.shared.windows.first?.makeKeyAndVisible()
-//        }
+        if #available(iOS 13.0, *) {
+            
+        } else {
+            window = UIWindow(frame: UIScreen.main.bounds)
+            let storageManager = StorageManager.shared
+
+            if storageManager.containsCoordinates {
+                if let coordinates = storageManager.storedCoordinates {
+                    let geoDataFinder = GeoDataFinder()
+                    geoDataFinder.delegate = self
+                    geoDataFinder.findBySpecified(coordinates: coordinates)
+                }
+            } else {
+                let storyboard = UIStoryboard(name: "Finder", bundle: Bundle.main)
+                let vc = storyboard.instantiateInitialViewController() as! FinderViewController
+                UIApplication.shared.windows.first?.rootViewController = vc
+                UIApplication.shared.windows.first?.makeKeyAndVisible()
+            }
+        }
         
         return true
     }
@@ -42,21 +46,18 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
     func application(_ application: UIApplication, didDiscardSceneSessions sceneSessions: Set<UISceneSession>) {
     }
 
-
 }
 
 
-//extension AppDelegate: GeoDataFinderDelegate {
-//    func didGeoDataFinded(with geoData: GeoData?) {
-//        let storyboard = UIStoryboard(name: "Weather", bundle: Bundle.main)
-//        DispatchQueue.main.async {
-//                    let vc = storyboard.instantiateInitialViewController() as! WeatherViewController
-//            vc.geoData = geoData
-//            UIApplication.shared.windows.first?.rootViewController = vc
-//            UIApplication.shared.windows.first?.makeKeyAndVisible()
-//        }
-//
-//    }
-//
-//}
+extension AppDelegate: GeoDataFinderDelegate {
+    func didGeoDataFinded(with geoData: GeoData?) {
+        let storyboard = UIStoryboard(name: "Weather", bundle: Bundle.main)
+        DispatchQueue.main.async {
+            let vc = storyboard.instantiateInitialViewController() as! WeatherViewController
+            vc.geoData = geoData
+            UIApplication.shared.windows.first?.rootViewController = vc
+            UIApplication.shared.windows.first?.makeKeyAndVisible()
+        }
+    }
+}
