--- WeatherApp/ru.lproj/Main.strings
@@ -1,1 +1,0 @@
-
